Add data source for database resources
This updates the scaffolding to include authentication for the provider,
along with a data source for retrieving details about a database.

I struggled to decide if it was better for me to just list out every
property of a database and explicitly configure each one individually,
or if I should leverage more of the SDK such that each property is
ultimately generated from the API. As a user of terraform, when
presented with the question of "should I unroll this loop?" - the answer
was almost always "yes" because if you don't you'd risk cascade
delete/recreates when an index gets off. But that was when using it, not
when writing a provider.

So I figured I should at least _try_ to use the generated SDK to
populate those properties. The downside is that if you look at this
provider you won't see an explicit enumeration of properties. The docs
will still be there with that information, but if you're someone digging
in to see the implementation details it will be an extra layer of
abstraction to wade through.

But the upside is that by using reflection, I am able to guarantee our
tf provider will be aligned with the API (so long as we are using the
correct version of the SDK for that api). It also means less code in the
tf provider itself. For this first datasource at least, it seems to be
working out alright - but we'll see with time if the generic definition
holds together.

DBAAS-7818