Add database resource for managing SkySQL database services
This implements the read/create/update/delete methods
for the resource, and includes the default id-based importer.

I also added a test for the resource which does a create with an example
resource, then updates the resource with a new name, and ultimately
deletes at the end.

We'll still want to implement a sweeper at some point to go purge those
records if something like that doesn't already exist in the backend.

DBAAS-7818