Update readme with information about SkySQL
The original readme from the scaffolding is mostly what we need, but I
needed to replace that blurb at the top with information about our
resources. I also mention the install make target, as that not being
mentioned tripped up Fares when he was trying it out a week or two ago.

We'll definitely want the technical writers to take a real pass at this.

DBAAS-7818