Add ssl_tls field for service resource and data_source
Now that we have the ability to enable/disable ssl encryption for a
service via the api, it needs to be exposed in the terraform provider as
well. Hopefully no one will ever use it to disable...

I also included a few updates from other changes in the api, including
the obfuscation of the `sys_` fields and change in the model types for
services.

DBAAS-8318