Add service credentials data source, block for service deployments
Now that we have an endpoint which provides the default credentials for
a new database, I added a new data source to use it. But that data
source doesn't really do any good unless the database has been deployed
and has a default user set up in it.

So, that meant I needed to introduce a blocking wait for the service
resource which would hang around until the database was deployed. To
help with testing, I added an option to skip this wait (so a test can
deploy and destroy immediately in dev where tasks aren't actually
ran). With that in place, we can then use the id of the resource in the
data source for its credential and that data source will wait for the
service to be ready. See the `main.tf` in the examples to see what I
mean (I use that file for impromptu manual testing, and it should show
how the dependency is represented).

Here's a reference doc for the blocking wait:
https://www.terraform.io/docs/extend/resources/retries-and-customizable-timeouts.html

DBAAS-8521