Include `required_providers` in provider doc
We'll need this so long as we're not in the terraform registry, so I
figure let's have it in the docs for now and we can remove it later if
it becomes unnecessary. This will also help anyone wanting to set up for
the first time and are looking at the docs (rather than directly at the
examples).